Cenrário 3

+++ src/trabalho/Escrita.java
@@ -15,19 +15,16 @@ public class Escrita {
     }
     
     //INÍCIO - PARTE 1 DO TRABALHO
-        //INÍCIO - CENÁRIO 1 E 2
+        //INÍCIO - CENÁRIO 1, 2 E 3
     public void imprimeCabecalio(int cenario, String execucao) {
         gravarSaida.write("-----------------------------------------------------------\n");
         gravarSaida.write("Parte 1 - Cenário "+cenario+" - "+execucao+"\n");
         gravarSaida.write("Tamanho\t\tTempo(ms)\tCópias\t\tComparações\n");
     }
-    public void imprimeDados(int parametro, long tempo, int copias, int comparacoes) {
+    public void imprimeDados(int parametro, long tempo, long copias, long comparacoes) {
         gravarSaida.write(parametro+"\t\t"+tempo+"\t\t"+copias+"\t\t"+comparacoes+"\n");
     }
-        //FINAL - CENÁRIO 1 E 2
-        //INÍCIO - CENÁRIO 3
-    
-        //FINAL - CENÁRIO 3
+        //FINAL - CENÁRIO 1, 2 E 3
         //INÍCIO - CENÁRIO 4
     
         //FINAL - CENÁRIO 4
